# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi untuk Service API Admin POS",
        "title": "Admin POS API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/master/customer/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create Master Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Customer"
                ],
                "summary": "Create Master Customer",
                "parameters": [
                    {
                        "description": "Create Master Customer",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemes.CustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses201Example"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses500Example"
                        }
                    }
                }
            }
        },
        "/api/v1/master/customer/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Delete Master Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Customer"
                ],
                "summary": "Get Delete Master Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Master Customer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses500Example"
                        }
                    }
                }
            }
        },
        "/api/v1/master/customer/results": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get List Master Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Customer"
                ],
                "summary": "Get List Master Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Use ASC or DESC | Available column sort : customer.id, customer.name, customer.phone, customer.active, merchant.id, merchant.name, outlet.id, outlet.name, customer.created_at, default is customer.created_at DESC | If you don't want to use it, fill it blank",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination, default is 1 | if you want to disable pagination, fill it with the number 0",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page for pagination, default is 10 | if you want to disable pagination, fill it with the number 0",
                        "name": "perpage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by merchant",
                        "name": "merchant_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by outlet",
                        "name": "outlet_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name using LIKE pattern",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.ResponsesPagination"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses500Example"
                        }
                    }
                }
            }
        },
        "/api/v1/master/customer/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Update Master Customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master Customer"
                ],
                "summary": "Get Update Master Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Update Master Customer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Update Master Customer",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemes.CustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses400Example"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses401Example"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses403Example"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses404Example"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses409Example"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemes.Responses500Example"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "schemes.CustomerRequest": {
            "type": "object",
            "required": [
                "address",
                "merchant_id",
                "name",
                "outlet_id",
                "phone"
            ],
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "address": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "JL. Pahlawan, Ngaliyan, Semarang"
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "customer tetap"
                },
                "merchant_id": {
                    "type": "string",
                    "example": "81c0b615-d575-4d30-a81a-6b8db70fd4e0"
                },
                "name": {
                    "type": "string",
                    "example": "candra"
                },
                "outlet_id": {
                    "type": "string",
                    "example": "4e769a02-0214-4277-90d0-bdf7f7b7c064"
                },
                "phone": {
                    "type": "string",
                    "minLength": 12,
                    "example": "087987875765"
                }
            }
        },
...
        "schemes.Responses": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status OK"
                }
            }
        },
        "schemes.Responses201Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 201
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Create Successfully"
                }
            }
        },
        "schemes.Responses400Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "schemes.Responses401Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "message": {
                    "type": "string",
                    "example": "Authorization header is required"
                }
            }
        },
        "schemes.Responses403Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status forbidden"
                }
            }
        },
        "schemes.Responses404Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status not found"
                }
            }
        },
        "schemes.Responses409Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 409
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status conflict data"
                }
            }
        },
        "schemes.Responses500Example": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status internal error"
                }
            }
        },
        "schemes.ResponsesPagination": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "status OK"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "example": 10
                },
                "total_data": {
                    "type": "integer",
                    "example": 50
                },
                "total_page": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "User JWT Bearer - Format Input Value : Bearer[ space ][ JWTToken ]",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}