# handlers/handler.customer.go
package handlers

import (
	"fmt"
	"math"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	gpc "github.com/restuwahyu13/go-playground-converter"

	"github.com/nutwreck/admin-pos-service/configs"
	"github.com/nutwreck/admin-pos-service/constants"
	"github.com/nutwreck/admin-pos-service/entities"
	"github.com/nutwreck/admin-pos-service/helpers"
	"github.com/nutwreck/admin-pos-service/pkg"
	"github.com/nutwreck/admin-pos-service/schemes"
)

type handleCustomer struct {
	customer entities.EntityCustomer
}

func NewHandlerCustomer(customer entities.EntityCustomer) *handleCustomer {
	return &handleCustomer{customer: customer}
}

/**
* ======================================
* Handler Ping Status Customer Teritory
*=======================================
 */

func (h *handleCustomer) HandlerPing(ctx *gin.Context) {
	helpers.APIResponse(ctx, "Ping Customer", http.StatusOK, nil)
}

/**
* =====================================
* Handler Create New Customer Teritory
*======================================
 */
// CreateMasterCustomer godoc
// @Summary		Create Master Customer
// @Description	Create Master Customer
// @Tags		Master Customer
// @Accept		json
// @Produce		json
// @Param		customer body schemes.CustomerRequest true "Create Master Customer"
// @Success 200 {object} schemes.Responses
// @Success 201 {object} schemes.Responses201Example
// @Failure 400 {object} schemes.Responses400Example
// @Failure 401 {object} schemes.Responses401Example
// @Failure 403 {object} schemes.Responses403Example
// @Failure 404 {object} schemes.Responses404Example
// @Failure 409 {object} schemes.Responses409Example
// @Failure 500 {object} schemes.Responses500Example
// @Security	ApiKeyAuth
// @Router /api/v1/master/customer/create [post]
func (h *handleCustomer) HandlerCreate(ctx *gin.Context) {
	var body schemes.Customer
	err := ctx.ShouldBindJSON(&body)

	if err != nil {
		helpers.APIResponse(ctx, "Parse json data from body failed", http.StatusBadRequest, nil)
		return
	}

	errors, code := ValidatorCustomer(ctx, body, "create")

	if code > 0 {
		helpers.ErrorResponse(ctx, errors)
		return
	}

	_, error := h.customer.EntityCreate(&body)

	if error.Type == "error_create_01" {
		helpers.APIResponse(ctx, "Customer phone number already taken", error.Code, nil)
		return
	}

	if error.Type == "error_create_02" {
		helpers.APIResponse(ctx, "Create new Customer failed", error.Code, nil)
		return
	}

	helpers.APIResponse(ctx, "Create new Customer successfully", http.StatusCreated, nil)
}

/**
* ======================================
* Handler Results All Customer Teritory
*=======================================
 */
// GetListMasterCustomer godoc
// @Summary		Get List Master Customer
// @Description	Get List Master Customer
// @Tags		Master Customer
// @Accept		json
// @Produce		json
// @Param sort query string false "Use ASC or DESC | Available column sort : customer.id, customer.name, customer.phone, customer.active, merchant.id, merchant.name, outlet.id, outlet.name, customer.created_at, default is customer.created_at DESC | If you don't want to use it, fill it blank"
// @Param page query int false "Page number for pagination, default is 1 | if you want to disable pagination, fill it with the number 0"
// @Param perpage query int false "Items per page for pagination, default is 10 | if you want to disable pagination, fill it with the number 0"
// @Param merchant_id query string false "Search by merchant"
// @Param outlet_id query string false "Search by outlet"
// @Param name query string false "Search by name using LIKE pattern"
// @Param id query string false "Search by ID"
// @Success 200 {object} schemes.ResponsesPagination
// @Failure 400 {object} schemes.Responses400Example
// @Failure 401 {object} schemes.Responses401Example
// @Failure 403 {object} schemes.Responses403Example
// @Failure 404 {object} schemes.Responses404Example
// @Failure 409 {object} schemes.Responses409Example
// @Failure 500 {object} schemes.Responses500Example
// @Security	ApiKeyAuth
// @Router /api/v1/master/customer/results [get]
func (h *handleCustomer) HandlerResults(ctx *gin.Context) {
	var (
		body          schemes.Customer
		reqPage       = configs.FirstPage
		reqPerPage    = configs.TotalPerPage
		pages         int
		perPages      int
		totalPagesDiv float64
		totalPages    int
		totalDatas    int
	)

	sortParam := ctx.DefaultQuery("sort", constants.EMPTY_VALUE)
	if sortParam != constants.EMPTY_VALUE {
		body.Sort = sortParam
	}
	pageParam := ctx.DefaultQuery("page", constants.EMPTY_VALUE)
	body.Page = reqPage
	if pageParam != constants.EMPTY_VALUE {
		page, err := strconv.Atoi(pageParam)
		if err != nil {
			helpers.APIResponsePagination(ctx, "Convert Params Failed", http.StatusInternalServerError, nil, pages, perPages, totalPages, totalDatas)
			return
		}
		reqPage = page
		body.Page = page
	}
	perPageParam := ctx.DefaultQuery("perpage", constants.EMPTY_VALUE)
	body.PerPage = reqPerPage
	if perPageParam != constants.EMPTY_VALUE {
		perPage, err := strconv.Atoi(perPageParam)
		if err != nil {
			helpers.APIResponsePagination(ctx, "Convert Params Failed", http.StatusInternalServerError, nil, pages, perPages, totalPages, totalDatas)
			return
		}
		reqPerPage = perPage
		body.PerPage = perPage
	}
	outletParam := ctx.DefaultQuery("outlet_id", constants.EMPTY_VALUE)
	if outletParam != constants.EMPTY_VALUE {
		body.OutletID = outletParam
	}
	merchantParam := ctx.DefaultQuery("merchant_id", constants.EMPTY_VALUE)
	if merchantParam != constants.EMPTY_VALUE {
		body.MerchantID = merchantParam
	}
	nameParam := ctx.DefaultQuery("name", constants.EMPTY_VALUE)
	if nameParam != constants.EMPTY_VALUE {
		body.Name = nameParam
	}
	idParam := ctx.DefaultQuery("id", constants.EMPTY_VALUE)
	if idParam != constants.EMPTY_VALUE {
		body.ID = idParam
	}

	res, totalData, error := h.customer.EntityResults(&body)

	if error.Type == "error_results_01" {
		helpers.APIResponsePagination(ctx, "Customer data not found", error.Code, nil, pages, perPages, totalPages, totalDatas)
		return
	}

	pages = reqPage
	perPages = reqPerPage
	if reqPerPage != 0 {
		totalPagesDiv = float64(totalData) / float64(reqPerPage)
	}
	totalPages = int(math.Ceil(totalPagesDiv))
	totalDatas = int(totalData)

	helpers.APIResponsePagination(ctx, "Customer data already to use", http.StatusOK, res, pages, perPages, totalPages, totalDatas)
}

/**
* ======================================
* Handler Result Customer By ID Teritory
*=======================================
 */

func (h *handleCustomer) HandlerResult(ctx *gin.Context) {
	var body schemes.Customer
	id := ctx.Param("id")
	body.ID = id

	errors, code := ValidatorCustomer(ctx, body, "result")

	if code > 0 {
		helpers.ErrorResponse(ctx, errors)
		return
	}

	res, error := h.customer.EntityResult(&body)

	if error.Type == "error_result_01" {
		helpers.APIResponse(ctx, fmt.Sprintf("Customer data not found for this id %s ", id), error.Code, nil)
		return
	}

	helpers.APIResponse(ctx, "Customer data already to use", http.StatusOK, res)
}

/**
* ======================================
* Handler Delete Customer By ID Teritory
*=======================================
 */
// GetDeleteMasterCustomer godoc
// @Summary		Get Delete Master Customer
// @Description	Get Delete Master Customer
// @Tags		Master Customer
// @Accept		json
// @Produce		json
// @Param		id query string true "Delete Master Customer"
// @Success 200 {object} schemes.Responses
// @Failure 400 {object} schemes.Responses400Example
// @Failure 401 {object} schemes.Responses401Example
// @Failure 403 {object} schemes.Responses403Example
// @Failure 404 {object} schemes.Responses404Example
// @Failure 409 {object} schemes.Responses409Example
// @Failure 500 {object} schemes.Responses500Example
// @Security	ApiKeyAuth
// @Router /api/v1/master/customer/delete [delete]
func (h *handleCustomer) HandlerDelete(ctx *gin.Context) {
	var body schemes.Customer
	id := ctx.DefaultQuery("id", constants.EMPTY_VALUE)
	body.ID = id

	errors, code := ValidatorCustomer(ctx, body, "delete")

	if code > 0 {
		helpers.ErrorResponse(ctx, errors)
		return
	}

	res, error := h.customer.EntityDelete(&body)

	if error.Type == "error_delete_01" {
		helpers.APIResponse(ctx, fmt.Sprintf("Customer data not found for this id %s ", id), error.Code, nil)
		return
	}

	if error.Type == "error_delete_02" {
		helpers.APIResponse(ctx, fmt.Sprintf("Delete Customer data for this id %v failed", id), error.Code, nil)
		return
	}

	helpers.APIResponse(ctx, fmt.Sprintf("Delete Customer data for this id %s success", id), http.StatusOK, res)
}

/**
* ======================================
* Handler Update Customer By ID Teritory
*=======================================
 */
// GetUpdateMasterCustomer godoc
// @Summary		Get Update Master Customer
// @Description	Get Update Master Customer
// @Tags		Master Customer
// @Accept		json
// @Produce		json
// @Param		id query string true "Update Master Customer"
// @Param		customer body schemes.CustomerRequest true "Update Master Customer"
// @Success 200 {object} schemes.Responses
// @Failure 400 {object} schemes.Responses400Example
// @Failure 401 {object} schemes.Responses401Example
// @Failure 403 {object} schemes.Responses403Example
// @Failure 404 {object} schemes.Responses404Example
// @Failure 409 {object} schemes.Responses409Example
// @Failure 500 {object} schemes.Responses500Example
// @Security	ApiKeyAuth
// @Router /api/v1/master/customer/update [put]
func (h *handleCustomer) HandlerUpdate(ctx *gin.Context) {
	var (
		body      schemes.Customer
		activeGet = false
	)
	id := ctx.DefaultQuery("id", constants.EMPTY_VALUE)
	body.ID = id
	body.Name = ctx.PostForm("name")
	body.Description = ctx.PostForm("description")
	body.Address = ctx.PostForm("address")
	body.Phone = ctx.PostForm("phone")
	body.MerchantID = ctx.PostForm("merchant_id")
	body.OutletID = ctx.PostForm("outlet_id")
	activeStr := ctx.PostForm("active")
	if activeStr == "true" {
		activeGet = constants.TRUE_VALUE
	}
	body.Active = &activeGet

	err := ctx.ShouldBindJSON(&body)

	if err != nil {
		helpers.APIResponse(ctx, "Parse json data from body failed", http.StatusBadRequest, nil)
		return
	}

	errors, code := ValidatorCustomer(ctx, body, "update")

	if code > 0 {
		helpers.ErrorResponse(ctx, errors)
		return
	}

	_, error := h.customer.EntityUpdate(&body)

	if error.Type == "error_update_01" {
		helpers.APIResponse(ctx, fmt.Sprintf("Customer data not found for this id %s ", id), error.Code, nil)
		return
	}

	if error.Type == "error_update_02" {
		helpers.APIResponse(ctx, fmt.Sprintf("Update Customer data failed for this id %s", id), error.Code, nil)
		return
	}

	helpers.APIResponse(ctx, fmt.Sprintf("Update Customer data success for this id %s", id), http.StatusCreated, nil)
}

/**
* ======================================
*  All Validator User Input For Customer
*=======================================
 */

func ValidatorCustomer(ctx *gin.Context, input schemes.Customer, Type string) (interface{}, int) {
	var schema gpc.ErrorConfig

	if Type == "create" {
		schema = gpc.ErrorConfig{
			Options: []gpc.ErrorMetaConfig{
				{
					Tag:     "required",
					Field:   "Name",
					Message: "Name is required on body",
				},
				{
					Tag:     "lowercase",
					Field:   "Name",
					Message: "Name must be lowercase",
				},
				{
					Tag:     "required",
					Field:   "Phone",
					Message: "Phone is required on body",
				},
				{
					Tag:     "numeric",
					Field:   "Phone",
					Message: "Phone must be number",
				},
				{
					Tag:     "gte",
					Field:   "Phone",
					Message: "Phone number must be 12 character",
				},
				{
					Tag:     "required",
					Field:   "Address",
					Message: "Address is required on body",
				},
				{
					Tag:     "max",
					Field:   "Address",
					Message: "Address maximal 1000 character",
				},
				{
					Tag:     "max",
					Field:   "Description",
					Message: "Description maximal 1000 character",
				},
				{
					Tag:     "required",
					Field:   "MerchantID",
					Message: "Merchant ID is required on body",
				},
				{
					Tag:     "uuid",
					Field:   "MerchantID",
					Message: "Merchant ID must be uuid",
				},
				{
					Tag:     "required",
					Field:   "OutletID",
					Message: "Outlet ID is required on body",
				},
				{
					Tag:     "uuid",
					Field:   "OutletID",
					Message: "Outlet ID must be uuid",
				},
			},
		}
	}

	if Type == "result" || Type == "delete" {
		schema = gpc.ErrorConfig{
			Options: []gpc.ErrorMetaConfig{
				{
					Tag:     "required",
					Field:   "ID",
					Message: "ID is required on param",
				},
				{
					Tag:     "uuid",
					Field:   "ID",
					Message: "ID must be uuid",
				},
			},
		}
	}

	if Type == "update" {
		schema = gpc.ErrorConfig{
			Options: []gpc.ErrorMetaConfig{
				{
					Tag:     "required",
					Field:   "ID",
					Message: "ID is required on param",
				},
				{
					Tag:     "uuid",
					Field:   "ID",
					Message: "ID must be uuid",
				},
				{
					Tag:     "required",
					Field:   "Name",
					Message: "Name is required on body",
				},
				{
					Tag:     "lowercase",
					Field:   "Name",
					Message: "Name must be lowercase",
				},
				{
					Tag:     "required",
					Field:   "Phone",
					Message: "Phone is required on body",
				},
				{
					Tag:     "numeric",
					Field:   "Phone",
					Message: "Phone must be number",
				},
				{
					Tag:     "gte",
					Field:   "Phone",
					Message: "Phone number must be 12 character",
				},
				{
					Tag:     "required",
					Field:   "Address",
					Message: "Address is required on body",
				},
				{
					Tag:     "max",
					Field:   "Address",
					Message: "Address maximal 1000 character",
				},
				{
					Tag:     "max",
					Field:   "Description",
					Message: "Description maximal 1000 character",
				},
				{
					Tag:     "required",
					Field:   "MerchantID",
					Message: "Merchant ID is required on body",
				},
				{
					Tag:     "uuid",
					Field:   "MerchantID",
					Message: "Merchant ID must be uuid",
				},
				{
					Tag:     "required",
					Field:   "OutletID",
					Message: "Outlet ID is required on body",
				},
				{
					Tag:     "uuid",
					Field:   "OutletID",
					Message: "Outlet ID must be uuid",
				},
			},
		}
	}

	err, code := pkg.GoValidator(&input, schema.Options)
	return err, code
}